Separating useNextPiecesSolo and Multi in 2 files, adding tests for solo, skipping multi
