+++ test/client/hooks/tetris-game/useNextPiecesMulti.test.js
@@ -1,12 +1,17 @@
 import React from "react";
-import { renderHook, act } from "@testing-library/react-hooks";
-import useNextPieces from "hooks/tetris-game/useNextPieces";
+import {
+  renderHook,
+  act
+} from "@testing-library/react-hooks";
+import useNextPiecesMulti from "hooks/tetris-game/useNextPiecesMulti";
 
-describe("useNextPieces", () => {
+describe("useNextPiecesMulti", () => {
   const useStateStub = jest.spyOn(React, "useState");
 
-  test("should create an interface to access nextPieces", () => {
-    const { result } = renderHook(useNextPieces);
+  test.skip("should create an interface to access nextPieces", () => {
+    const {
+      result
+    } = renderHook(useNextPiecesMulti);
 
     expect(useStateStub).toHaveBeenCalledTimes(1);
     expect(result.current.nextPieces).toHaveLength(3);
